+++ src/utils/index.js
@@ -5,7 +5,9 @@ module.exports = {
   ...require('./globalConfig'),
   ...require('./getters'),
   ...require('./helpers'),
+  ...require('./process'),
   ...require('./log'),
+  ...require('./network'),
   ...require('./package'),
   ...require('./template'),
   ...require('./task'),
