+++ src/utils/network/index.js
@@ -1,0 +1,4 @@
+module.exports = {
+  ...require('./getProcessesUsingPort'),
+  ...require('./printProcesses')
+}
