+++ src/utils/process/index.js
@@ -1,0 +1,3 @@
+module.exports = {
+  ...require('./kill')
+}
